requirements.txt
flask
flask-cors
colorama
yt-dlp
python-ffmpeg
pystray
Pillow
appdirs
requests
packaging